Move fmr module below api module
