+++ docs/howto/config.rst
@@ -107,7 +107,7 @@ the endpoint is the URL at which the FMR is available, followed by
 
 .. code-block:: python
 
-    from pysdmx.fmr import AsyncRegistryClient
+    from pysdmx.api.fmr import AsyncRegistryClient
 
     client = AsyncRegistryClient("[endpoint_comes_here]")
