+++ src/pysdmx/api/fmr/__init__.py
@@ -14,8 +14,8 @@ from typing import (
 import httpx
 from msgspec.json import decode
 
+from pysdmx.api.fmr.reader import Deserializer
 from pysdmx.errors import ClientError, NotFound, ServiceError, Unavailable
-from pysdmx.fmr.reader import Deserializer
 from pysdmx.io.json.fusion.reader import deserializers as fusion_readers
 from pysdmx.io.json.sdmxjson2.reader import deserializers as sdmx_readers
 from pysdmx.model import (
