+++ src/pysdmx/io/json/fusion/messages/report.py
@@ -4,7 +4,7 @@ from typing import Sequence
 
 from msgspec import Struct
 
-from pysdmx.fmr.reader import _merge_attributes
+from pysdmx.api.fmr.reader import _merge_attributes
 from pysdmx.io.json.fusion.messages.core import FusionString
 from pysdmx.model import MetadataAttribute, MetadataReport
