+++ tests/fmr/code_map_checks.py
@@ -4,7 +4,7 @@ from typing import Pattern
 import httpx
 import pytest
 
-from pysdmx.fmr import AsyncRegistryClient, RegistryClient
+from pysdmx.api.fmr import AsyncRegistryClient, RegistryClient
 from pysdmx.model.concept import DataType
 from pysdmx.model.map import (
     MultiRepresentationMap,
