+++ tests/fmr/fusion/test_providers.py
@@ -1,7 +1,7 @@
 import pytest
 import tests.fmr.provider_checks as checks
 
-from pysdmx.fmr import AsyncRegistryClient, Format, RegistryClient
+from pysdmx.api.fmr import AsyncRegistryClient, Format, RegistryClient
 
 
 @pytest.fixture()
