+++ tests/fmr/fusion/test_vl_codes.py
@@ -1,7 +1,7 @@
 import pytest
 import tests.fmr.vl_code_checks as checks
 
-from pysdmx.fmr import AsyncRegistryClient, Format, RegistryClient
+from pysdmx.api.fmr import AsyncRegistryClient, Format, RegistryClient
 
 
 @pytest.fixture()
