+++ tests/fmr/mult_reports_checks.py
@@ -2,7 +2,7 @@ from typing import Sequence
 
 import httpx
 
-from pysdmx.fmr import AsyncRegistryClient, RegistryClient
+from pysdmx.api.fmr import AsyncRegistryClient, RegistryClient
 from pysdmx.model import MetadataAttribute
