+++ tests/fmr/report_checks.py
@@ -1,7 +1,7 @@
 import httpx
 import pytest
 
-from pysdmx.fmr import AsyncRegistryClient, RegistryClient
+from pysdmx.api.fmr import AsyncRegistryClient, RegistryClient
 from pysdmx.model import MetadataAttribute, MetadataReport
