+++ tests/fmr/vl_code_checks.py
@@ -1,6 +1,6 @@
 import httpx
 
-from pysdmx.fmr import AsyncRegistryClient, RegistryClient
+from pysdmx.api.fmr import AsyncRegistryClient, RegistryClient
 from pysdmx.model import Code, Codelist
